Principal Component Analysis Classifier
This notebook shows how PCA can be applied to reduce dimensionality of a high dimensional dataset MNIST which is dataset of handwritten digits between 0 to 9 where each datapoint is 8x8 image of integer pixels in the range 0 to 16 & can be represented as a vector of 64 dimensions. sklearn MNIST have 5620 images. More details of MNIST can be found on this [link](https://scikit-learn.org/stable/datasets/index.html#optical-recognition-of-handwritten-digits-dataset).